requirements.txt
requests
pillow
diffsync
pyyaml
jinja2
